+++ packages/perseus/src/perseus-types.ts
@@ -724,6 +724,7 @@ export type LockedPointType = {
     coord: Coord;
     color: LockedFigureColor;
     filled: boolean;
+    labels: LockedLabelType[];
 };
 
 export type LockedLineType = {
@@ -734,6 +735,7 @@ export type LockedLineType = {
     lineStyle: LockedLineStyle;
     showPoint1: boolean;
     showPoint2: boolean;
+    labels: LockedLabelType[];
 };
 
 export type LockedVectorType = {