+++ packages/perseus/src/widgets/interactive-graphs/stateful-mafs-graph.tsx
@@ -15,9 +15,10 @@ import {getGradableGraph, getRadius} from "./reducer/interactive-graph-state";
 import type {InteractiveGraphProps, InteractiveGraphState} from "./types";
 import type {PerseusGraphType} from "../../perseus-types";
 import type {Widget} from "../../renderer";
+import type {APIOptions} from "../../types";
 
 export type StatefulMafsGraphProps = {
-    showLabelsFlag?: boolean;
+    flags?: APIOptions["flags"];
     box: [number, number];
     backgroundImage?: InteractiveGraphProps["backgroundImage"];
     graph: PerseusGraphType;